Make Customer Finder Configurable
Allow the distance provider to be configurable so if users want to use
a different algorithm or gem to calculate the distance, they can.

Allow the file path to the customers be configurable so different
customer files can be used.
