+++ Gemfile
@@ -2,3 +2,4 @@ source 'https://rubygems.org'
 
 gem 'rspec', '~> 3.8'
 gem 'dry-struct', '~> 0.7'
+gem 'geokit', '~> 1.13'
