+++ bin/console
@@ -1,5 +1,6 @@
 #!/usr/bin/env ruby
 
+require 'bundler/setup'
 require './lib/customer_finder'
 
 require 'irb'
